+++ R/dummy.R
@@ -267,7 +267,6 @@ bake.step_dummy <- function(object, new_data, ...) {
   }
 
   col_names <- names(object$levels)
-  keep_original_cols <- get_keep_original_cols(object)
 
   ## `na.action` cannot be passed to `model.matrix` but we
   ## can change it globally for a bit
@@ -335,10 +334,10 @@ bake.step_dummy <- function(object, new_data, ...) {
     indicators <- check_name(indicators, new_data, object, names(indicators))
 
     new_data <- vec_cbind(new_data, indicators)
-    if (any(!object$preserve, !keep_original_cols)) {
-      new_data[[col_names[i]]] <- NULL
-    }
   }
+
+  new_data <- remove_original_cols(new_data, object, col_names)
+
   new_data
 }