+++ R/harmonic.R
@@ -318,11 +318,8 @@ bake.step_harmonic <- function(object, new_data, ...) {
     new_data <- vec_cbind(new_data, res)
   }
 
-  keep_original_cols <- get_keep_original_cols(object)
-  if (!keep_original_cols) {
-    new_data <-
-      new_data[, !(colnames(new_data) %in% col_names), drop = FALSE]
-  }
+  new_data <- remove_original_cols(new_data, object, col_names)
+
   new_data
 }