+++ R/kpca_rbf.R
@@ -161,11 +161,7 @@ bake.step_kpca_rbf <- function(object, new_data, ...) {
     comps <- as_tibble(comps)
     comps <- check_name(comps, new_data, object)
     new_data <- vec_cbind(new_data, comps)
-    keep_original_cols <- get_keep_original_cols(object)
-
-    if (!keep_original_cols) {
-      new_data <- new_data[, !(colnames(new_data) %in% object$columns), drop = FALSE]
-    }
+    new_data <- remove_original_cols(new_data, object, object$columns)
   }
   new_data
 }