+++ R/time.R
@@ -144,10 +144,7 @@ bake.step_time <- function(object, new_data, ...) {
     new_data <- vec_cbind(new_data, time_values)
   }
 
-  keep_original_cols <- get_keep_original_cols(object)
-  if (!keep_original_cols) {
-    new_data <- new_data[, !(colnames(new_data) %in% object$columns), drop = FALSE]
-  }
+  new_data <- remove_original_cols(new_data, object, object$columns)
 
   new_data
 }